--- src/client/SessionStart.java
@@ -1,56 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package client;
-
-import java.awt.Window;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import javax.swing.ImageIcon;
-import javax.swing.JButton;
-import javax.swing.JDialog;
-import javax.swing.JOptionPane;
-import javax.swing.JPanel;
-
-/**
- *
- * @author admin
- */
-public class SessionStart extends JPanel implements ActionListener{
-   
-    public static String selection = "";
-    
-    public SessionStart (){
-        JButton log = new JButton("Login");        
-        JButton reg = new JButton("Register");        
-        JButton guest = new JButton("Guesst");
-        log.addActionListener(this);
-        reg.addActionListener(this);
-        guest.addActionListener(this);
-        add(log);
-        add(reg);
-        add(guest);
-    }
-    
-    @Override
-    public void actionPerformed(ActionEvent e) {
-        selection = e.getActionCommand();
-        System.out.println(selection);
-        Window[] windows = Window.getWindows();
-                for (Window window : windows) {
-                    if (window instanceof JDialog) {
-                        JDialog dialog = (JDialog) window;
-                        if (dialog.getContentPane().getComponentCount() == 1
-                            && dialog.getContentPane().getComponent(0) instanceof JOptionPane){
-                            dialog.dispose();
-                        }
-                    }
-                }
-    }
-    
-    public static String getSelection(){
-        return selection;
-    }
-    
-}
