+++ src/client/helpers/GetIP.java
@@ -1,4 +1,4 @@
-package client;
+package client.helpers;
 
 /*
  * To change this license header, choose License Headers in Project Properties.
